Bot/requirements.txt
aiogram==2.11.2
asyncio-nats-client==0.11.4
python-dotenv==0.15.0
requests==2.25.1
pymongo==3.11.3
motor==2.4.0